requirement.txt
fastapi
uvicorn
SQLAlchemy
PyMySQL
cryptography
mysql-connector-python
python-dotenv
pydantic_settings
asyncmy